+++ CHANGELOG.md
@@ -36,6 +36,7 @@ New features:
 * core env: working lxd prototype
 
 Fix:
+* file.stat: handle empty stdout
 * file.hash: enrich error message
 * file.download: honors sudo
 * service.init: disable templating if no context